10-7959a | Duty to Assist page config
