+++ src/applications/ivc-champva/10-7959a/chapters/resubmission/supportingDocOptions.js
@@ -23,9 +23,7 @@ export default {
       title: INPUT_LABEL,
       labels: SCHEMA_LABELS,
     }),
-    'view:addtlInfo': {
-      ...descriptionUI(ResubmissionOptionsAddtlInfo),
-    },
+    'view:addtlInfo': descriptionUI(ResubmissionOptionsAddtlInfo),
   },
   schema: {
     type: 'object',
